Navigations refactored
